+++ src/MicroElements.Metadata/Metadata/DynamicContainer.cs
@@ -35,8 +35,11 @@ namespace MicroElements.Metadata
         /// <inheritdoc />
         public override bool TryGetMember(GetMemberBinder binder, out object result)
         {
-            var propertyValue = _propertyContainer.GetPropertyValueUntyped(
-                new SearchCondition(binder.Name, ignoreCase: _ignoreCase, searchInParent: _searchInParent, returnNotDefined: false));
+            var propertyValue = _propertyContainer.GetPropertyValueUntyped(Search
+                .ByNameOrAlias(binder.Name, _ignoreCase)
+                .SearchInParent(_searchInParent)
+                .ReturnNull());
+
             if (propertyValue != null)
             {
                 result = propertyValue.ValueUntyped;