+++ src/MicroElements.Metadata/Metadata/PropertyContainerExtensions.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
+using System.Collections.Generic;
 using System.Linq;
 using MicroElements.Functional;
 
@@ -13,53 +14,33 @@ namespace MicroElements.Metadata
     public static partial class PropertyContainerExtensions
     {
         /// <summary>
-        /// Gets property and value by search conditions.
+        /// Searches property and value by search conditions.
         /// </summary>
         /// <param name="propertyContainer">Property container.</param>
+        /// <param name="property">Property to search.</param>
         /// <param name="search">Search conditions.</param>
         /// <returns><see cref="IPropertyValue"/> or null.</returns>
-        public static IPropertyValue GetPropertyValueUntyped(this IPropertyContainer propertyContainer, SearchCondition search)
+        public static IPropertyValue SearchPropertyValueUntyped(this IPropertyContainer propertyContainer, IProperty property, SearchOptions search = default)
         {
-            // Ищем в своих свойствах.
-            IPropertyValue propertyValue;
-            if (search.Property != null)
-            {
-                // Search By Property
-                propertyValue = propertyContainer.FirstOrDefault(pv => pv.PropertyUntyped == search.Property);
-            }
-            else
-            {
-                // Search By Name or Alias
-                propertyValue = propertyContainer.FirstOrDefault(pv => pv.IsMatchesByNameOrAlias(search.Name, search.IgnoreCase));
-            }
+            propertyContainer.AssertArgumentNotNull(nameof(propertyContainer));
+            property.AssertArgumentNotNull(nameof(property));
+
+            // Search property by EqualityComparer
+            IPropertyValue propertyValue = propertyContainer.FirstOrDefault(pv => search.PropertyComparer.Equals(pv.PropertyUntyped, property));
 
             if (propertyValue != null)
                 return propertyValue;
 
-            // Ищем у родителя
+            // Search in parent
             if (search.SearchInParent && propertyContainer.ParentSource != null)
             {
-                propertyValue = propertyContainer.ParentSource.GetPropertyValueUntyped(search);
+                propertyValue = propertyContainer.ParentSource.SearchPropertyValueUntyped(property, search);
 
                 if (propertyValue != null)
                     return propertyValue;
             }
 
-            return search.ReturnNotDefined ? PropertyValue.Create(search.Property ?? new Property<string>(search.Name), null, ValueSource.NotDefined) : null;
-        }
-
-        /// <summary>
-        /// Gets property and value by search conditions.
-        /// </summary>
-        /// <param name="propertyContainer">Property container.</param>
-        /// <param name="propertyName">Search conditions.</param>
-        /// <param name="searchInParent">Search in parent.</param>
-        /// <returns>value or null.</returns>
-        public static object GetValueUntyped(this IPropertyContainer propertyContainer, string propertyName, bool searchInParent = true)
-        {
-            return propertyContainer
-                .GetPropertyValueUntyped(new SearchCondition(propertyName, ignoreCase: true, searchInParent: searchInParent, returnNotDefined: false))
-                ?.ValueUntyped;
+            return search.ReturnNotDefined ? PropertyValue.Create(property, property.Type.GetDefaultValue(), ValueSource.NotDefined) : null;
         }
 
         /// <summary>
@@ -67,10 +48,127 @@ namespace MicroElements.Metadata
         /// </summary>
         /// <typeparam name="T">Property type.</typeparam>
         /// <param name="propertyContainer">Property container.</param>
+        /// <param name="property">Property to search.</param>
         /// <param name="search">Search conditions.</param>
         /// <returns><see cref="IPropertyValue"/> or null.</returns>
-        public static IPropertyValue<T> GetPropertyValue<T>(this IPropertyContainer propertyContainer, SearchCondition search) =>
-            (IPropertyValue<T>)propertyContainer.GetPropertyValueUntyped(search);
+        public static IPropertyValue<T> GetPropertyValue<T>(this IPropertyContainer propertyContainer, IProperty<T> property, SearchOptions search = default)
+        {
+            propertyContainer.AssertArgumentNotNull(nameof(propertyContainer));
+            property.AssertArgumentNotNull(nameof(property));
+
+            // Base search
+            IPropertyValue propertyValue = propertyContainer
+                .SearchPropertyValueUntyped(property, search
+                    .UseDefaultValue(false)
+                    .ReturnNull());
+
+            // Good job - return result!
+            if (propertyValue != null)
+                return (IPropertyValue<T>)propertyValue;
+
+            // Property can be calculated.
+            if (search.CalculateValue && property.Calculate != null)
+            {
+                var calculatedValue = property.Calculate(propertyContainer);
+                return new PropertyValue<T>(property, calculatedValue, ValueSource.Calculated);
+            }
+
+            // Maybe default value?
+            if (search.UseDefaultValue)
+                return new PropertyValue<T>(property, property.DefaultValue(), ValueSource.DefaultValue);
+
+            // Return null or NotDefined
+            return search.ReturnNotDefined ? new PropertyValue<T>(property, default, ValueSource.NotDefined) : null;
+        }
+
+        /// <summary>
+        /// Calls <see cref="GetPropertyValue{T}(IPropertyContainer,SearchOptions)"/> using type from <paramref name="property"/>.
+        /// </summary>
+        /// <param name="propertyContainer">Property container.</param>
+        /// <param name="property">Property to search.</param>
+        /// <param name="search">Search conditions.</param>
+        /// <returns><see cref="IPropertyValue"/> or null.</returns>
+        public static IPropertyValue GetPropertyValueCompiled(this IPropertyContainer propertyContainer, IProperty property, SearchOptions search)
+        {
+            static IPropertyValue GetPropertyValueAdapter<T>(IPropertyContainer propertyContainer, IProperty property, SearchOptions search) => GetPropertyValue(propertyContainer, (IProperty<T>)property, search);
+            var getPropertyValueCompiled = CodeCompiler.CachedCompiledFunc<IPropertyContainer, IProperty, SearchOptions, IPropertyValue>(property.Type, GetPropertyValueAdapter<CodeCompiler.GenericType>);
+            return getPropertyValueCompiled(propertyContainer, property, search);
+        }
+
+        /// <summary>
+        /// Gets property and value by search conditions.
+        /// </summary>
+        /// <param name="propertyContainer">Property container.</param>
+        /// <param name="property">Property to search.</param>
+        /// <param name="search">Search conditions.</param>
+        /// <returns><see cref="IPropertyValue"/> or null.</returns>
+        public static IPropertyValue GetPropertyValueUntyped(this IPropertyContainer propertyContainer, IProperty property, SearchOptions search = default)
+        {
+            propertyContainer.AssertArgumentNotNull(nameof(propertyContainer));
+            property.AssertArgumentNotNull(nameof(property));
+
+            if (search.CanUseUntypedSearch() || property.Type == typeof(Search.UntypedSearch))
+            {
+                return propertyContainer.SearchPropertyValueUntyped(property, search);
+            }
+
+            return propertyContainer.GetPropertyValueCompiled(property, search);
+        }
+
+        /// <summary>
+        /// Returns true if untyped search can be used.
+        /// </summary>
+        /// <param name="search">SearchOptions.</param>
+        /// <returns>true if untyped search can be used.</returns>
+        public static bool CanUseUntypedSearch(this in SearchOptions search) => search.CalculateValue == false && search.UseDefaultValue == false;
+
+        /// <summary>
+        /// Searches property and value by search conditions.
+        /// NOTE: PropertyToSearch must be set in <paramref name="search"/>.
+        /// </summary>
+        /// <param name="propertyContainer">Property container.</param>
+        /// <param name="search">Search conditions.</param>
+        /// <returns><see cref="IPropertyValue"/> or null.</returns>
+        public static IPropertyValue SearchPropertyValueUntyped(this IPropertyContainer propertyContainer, SearchOptions search)
+        {
+            if (search.SearchProperty == null)
+                throw new InvalidOperationException("SearchProperty must be set in SearchOptions.");
+
+            return propertyContainer.SearchPropertyValueUntyped(search.SearchProperty, search);
+        }
+
+        /// <summary>
+        /// Gets property and value by search conditions.
+        /// NOTE: PropertyToSearch must be set in <paramref name="search"/>.
+        /// </summary>
+        /// <param name="propertyContainer">Property container.</param>
+        /// <param name="search">Search conditions.</param>
+        /// <returns><see cref="IPropertyValue"/> or null.</returns>
+        public static IPropertyValue GetPropertyValueUntyped(this IPropertyContainer propertyContainer, SearchOptions search)
+        {
+            if (search.SearchProperty == null)
+                throw new InvalidOperationException("SearchProperty must be set in SearchOptions.");
+
+            return propertyContainer.GetPropertyValueUntyped(search.SearchProperty, search);
+        }
+
+        /// <summary>
+        /// Gets property and value by search conditions.
+        /// NOTE: PropertyToSearch must be set in <paramref name="search"/>.
+        /// </summary>
+        /// <typeparam name="T">Property type.</typeparam>
+        /// <param name="propertyContainer">Property container.</param>
+        /// <param name="search">Search conditions.</param>
+        /// <returns><see cref="IPropertyValue"/> or null.</returns>
+        public static IPropertyValue<T> GetPropertyValue<T>(this IPropertyContainer propertyContainer, SearchOptions search)
+        {
+            if (search.SearchProperty == null)
+                throw new InvalidOperationException("SearchProperty must be set in SearchOptions.");
+            if (search.SearchProperty.Type != typeof(T))
+                throw new InvalidOperationException($"SearchProperty type must should be {typeof(T)} but is {search.SearchProperty.Type}.");
+
+            return propertyContainer.GetPropertyValue((IProperty<T>)search.SearchProperty, search);
+        }
 
         /// <summary>
         /// Gets property and value by property.
@@ -82,6 +180,7 @@ namespace MicroElements.Metadata
         /// <param name="calculateValue">Calculate value if value was not found.</param>
         /// <param name="useDefaultValue">Use default value from property.</param>
         /// <param name="returnNotDefined">Return not null PropertyValue with <see cref="ValueSource.NotDefined"/> if no PropertyValue was found.</param>
+        /// <param name="propertyComparer">Property equality comparer.</param>
         /// <returns><see cref="IPropertyValue"/> or null.</returns>
         public static IPropertyValue<T> GetPropertyValue<T>(
             this IPropertyContainer propertyContainer,
@@ -89,26 +188,16 @@ namespace MicroElements.Metadata
             bool searchInParent = true,
             bool calculateValue = true,
             bool useDefaultValue = true,
-            bool returnNotDefined = true)
+            bool returnNotDefined = true,
+            IEqualityComparer<IProperty> propertyComparer = null)
         {
-            IPropertyValue<T> propertyValue = (IPropertyValue<T>)propertyContainer.GetPropertyValueUntyped(
-                new SearchCondition(property, searchInParent: searchInParent, returnNotDefined: false));
-
-            if (propertyValue.HasValue())
-                return propertyValue;
-
-            // Свойство не в списке свойств, но может его можно вычислить.
-            if (calculateValue && property.Calculate != null)
-            {
-                var calculatedValue = property.Calculate(propertyContainer);
-                return new PropertyValue<T>(property, calculatedValue, ValueSource.Calculated);
-            }
-
-            // Вернем значение по умолчанию.
-            if (useDefaultValue)
-                return new PropertyValue<T>(property, property.DefaultValue(), ValueSource.DefaultValue);
-
-            return returnNotDefined ? new PropertyValue<T>(property, default, ValueSource.NotDefined) : null;
+            return propertyContainer.GetPropertyValue(
+                property, new SearchOptions(
+                    propertyComparer: propertyComparer,
+                    searchInParent: searchInParent,
+                    calculateValue: calculateValue,
+                    useDefaultValue: useDefaultValue,
+                    returnNotDefined: returnNotDefined));
         }
 
         /// <summary>
@@ -118,8 +207,25 @@ namespace MicroElements.Metadata
         /// <param name="property">Property to search.</param>
         /// <param name="searchInParent">Search in parent.</param>
         /// <returns><see cref="IPropertyValue"/> or null.</returns>
-        public static IPropertyValue GetPropertyValueUntyped(this IPropertyContainer propertyContainer, IProperty property, bool searchInParent = true) =>
-            propertyContainer.GetPropertyValueUntyped(new SearchCondition(property, searchInParent: searchInParent, returnNotDefined: true));
+        public static IPropertyValue GetPropertyValueUntyped(this IPropertyContainer propertyContainer, IProperty property, bool searchInParent = true)
+            => propertyContainer.GetPropertyValueUntyped(property, new SearchOptions(searchInParent: searchInParent, returnNotDefined: true));
+
+        /// <summary>
+        /// Gets property and value by search conditions.
+        /// </summary>
+        /// <param name="propertyContainer">Property container.</param>
+        /// <param name="propertyName">Search conditions.</param>
+        /// <param name="searchInParent">Search in parent.</param>
+        /// <returns>value or null.</returns>
+        public static object GetValueUntyped(this IPropertyContainer propertyContainer, string propertyName, bool searchInParent = true)
+        {
+            IPropertyValue propertyValue = propertyContainer.GetPropertyValueUntyped(Search
+                .ByNameOrAlias(propertyName, ignoreCase: true)
+                .SearchInParent(searchInParent)
+                .ReturnNull());
+
+            return propertyValue?.ValueUntyped;
+        }
 
         /// <summary>
         /// Returns true if <paramref name="propertyContainer"/> has property.
@@ -127,10 +233,10 @@ namespace MicroElements.Metadata
         /// </summary>
         /// <param name="propertyContainer">Property container.</param>
         /// <param name="property">Property to search.</param>
-        /// <param name="searchInParent">Search in parent.</param>
+        /// <param name="search">Search conditions.</param>
         /// <returns>True if property exists in container.</returns>
-        public static bool HasValue(this IPropertyContainer propertyContainer, IProperty property, bool searchInParent = true) =>
-            propertyContainer.GetPropertyValueUntyped(property, searchInParent).HasValue();
+        public static bool HasValue(this IPropertyContainer propertyContainer, IProperty property, SearchOptions search) =>
+            propertyContainer.GetPropertyValueUntyped(property, search).HasValue();
 
         /// <summary>
         /// Sets property value if property is not set.
@@ -141,93 +247,9 @@ namespace MicroElements.Metadata
         /// <param name="value">Value to set.</param>
         public static void SetValueIfNotSet<T>(this IMutablePropertyContainer propertyContainer, IProperty<T> property, T value)
         {
-            var propertyValue = propertyContainer.GetPropertyValueUntyped(property, searchInParent: false);
+            var propertyValue = propertyContainer.GetPropertyValueUntyped(property, SearchOptions.Default.SearchInParent(false).ReturnNull());
             if (propertyValue.IsNullOrNotDefined())
                 propertyContainer.SetValue(property, value);
-        }
-    }
-
-    /// <summary>
-    /// Search options.
-    /// </summary>
-    public readonly struct SearchCondition
-    {
-        /// <summary>
-        /// Property to search.
-        /// </summary>
-        public readonly IProperty Property;
-
-        /// <summary>
-        /// Search by name and alias.
-        /// </summary>
-        public readonly string Name;
-
-        /// <summary>
-        /// Name search ignore case.
-        /// </summary>
-        public readonly bool IgnoreCase;
-
-        /// <summary>
-        /// >Do search in parent.
-        /// </summary>
-        public readonly bool SearchInParent;
-
-        /// <summary>
-        /// Return not null PropertyValue with <see cref="ValueSource.NotDefined"/> if no PropertyValue was found.
-        /// </summary>
-        public readonly bool ReturnNotDefined;
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="SearchCondition"/> struct.
-        /// </summary>
-        /// <param name="property">Search by property.</param>
-        /// <param name="searchInParent">Do search in parent.</param>
-        /// <param name="returnNotDefined">Return not null PropertyValue with <see cref="ValueSource.NotDefined"/> if no PropertyValue was found.</param>
-        public SearchCondition(IProperty property, bool searchInParent, bool returnNotDefined = true)
-        {
-            Property = property.AssertArgumentNotNull(nameof(property));
-
-            Name = null;
-            IgnoreCase = false;
-
-            SearchInParent = searchInParent;
-            ReturnNotDefined = returnNotDefined;
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="SearchCondition"/> struct.
-        /// Search by name.
-        /// </summary>
-        /// <param name="name">Search by name.</param>
-        /// <param name="ignoreCase">Comparison.</param>
-        /// <param name="searchInParent">Do search in parent.</param>
-        /// <param name="returnNotDefined">Return not null PropertyValue with <see cref="ValueSource.NotDefined"/> if no PropertyValue was found.</param>
-        public SearchCondition(string name, bool ignoreCase, bool searchInParent, bool returnNotDefined = true)
-        {
-            Property = null;
-
-            Name = name;
-            IgnoreCase = ignoreCase;
-
-            SearchInParent = searchInParent;
-            ReturnNotDefined = returnNotDefined;
-        }
-    }
-
-    /// <summary>
-    /// Extension methods for search.
-    /// </summary>
-    public static class SearchExtensions
-    {
-        public static bool IsMatchesByNameOrAlias(this IPropertyValue propertyValue, string propertyName, bool ignoreCase = false)
-        {
-            StringComparison stringComparison = ignoreCase ? StringComparison.OrdinalIgnoreCase : StringComparison.Ordinal;
-
-            if (propertyName == null)
-                return false;
-
-            return propertyName.Equals(propertyValue.PropertyUntyped.Name, stringComparison)
-                   || propertyName.Equals(propertyValue.PropertyUntyped.Alias, stringComparison);
         }
     }
 }